Armstrong Number
Here is code, in Java, to determine if a number is Armstrong. Included within the code, in the header, is a brief description of how the program runs. If you are unaware, a number is Armstrong if the sum of all of it's digits individually raised to the power of the number of digits is equal to the original number.