Assignment 4
In this Assignment, we were tasked with using Linear, and Binary searches on a selection of parallel arrays. Additionally, we were asked to sort a randomly generated array of 10 integers, all between 1 - 500. The purpose was to see how that various searching forms could be utilized for a more efficient sort, so I also added a timer to determine how long the execution of one particular search/sort took.